+++ antifraud/src/main/java/com/bank/antifraud/dto/SuspiciousCardTransferDto.java
@@ -1,9 +1,7 @@
 package com.bank.antifraud.dto;
 
-import java.math.BigInteger;
-
 public record SuspiciousCardTransferDto
-        (BigInteger id,
+        (Long id,
          Long cardTransferId,
          Boolean isBlocked,
          Boolean isSuspicious,
